@@ -7,17 +7,6 @@
     "type": "git",
     "url": "git+https://github.com/kfaubel/khoto.git"
   },
-  "dependencies": {
-    "axios": "^0.21.1",
-    "bootstrap": "^5.0.1",
-    "core-js": "^3.13.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.6.0",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "react-select": "^4.3.1"
-  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
@@ -25,6 +14,19 @@
     "eject": "react-scripts eject",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build"
+  },
+  "dependencies": {
+    "axios": "^1.4.0",
+    "bootstrap": "^5.3.1",
+    "buffer": "^6.0.3",
+    "core-js": "^3.32.1",
+    "crypto": "npm:crypto-browserify",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.8.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.15.0",
+    "react-select": "^5.7.4",
+    "stream": "npm:stream-browserify"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -42,10 +44,7 @@
     ]
   },
   "devDependencies": {
-    "@babel/plugin-bugfix-v8-spread-parameters-in-optional-chaining": "^7.13.12",
-    "babel-core": "^6.26.3",
-    "gh-pages": "^3.2.0",
-    "typescript": "^4.3.2",
-    "react-scripts": "^4.0.3"
+    "gh-pages": "^6.0.0",
+    "react-scripts": "^5.0.1"
   }
 }
